+++ test/Users/index.test.js
@@ -1,4 +1,4 @@
-import { testGetter, testPut } from "../utils/http";
+import { testGetter, testPost, testPut } from "../utils/http";
 import Users from "../../src/Users";
 
 describe("Users#self", () => {
@@ -21,6 +21,18 @@ describe("Users#updateAccount", () => {
       return new Users(client).updateAccount({
         company: "New company",
       });
+    }
+  );
+});
+
+describe("Users#stopFreeTrial", () => {
+  testPost(
+    "https://auth.scalingo.com/v1/users/stop_free_trial",
+    null,
+    null,
+    null,
+    (client) => {
+      return new Users(client).stopFreeTrial();
     }
   );
 });